% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmstatrExt-package.R
\docType{package}
\name{cmstatrExt-package}
\alias{cmstatrExt}
\alias{cmstatrExt-package}
\title{cmstatrExt: More Statistical Methods for Composite Material Data}
\description{
\code{cmstatrExt} is a package that contains statistical methods that can be
used for composite material data, but that are not contained within
CMH-17.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmstatr/cmstatrExt}
  \item \url{https://cmstatrExt.cmstatr.net}
  \item Report bugs at \url{https://github.com/cmstatr/cmstatrExt/issues}
}

}
\author{
\strong{Maintainer}: Stefan Kloppenborg \email{stefan@kloppenborg.ca}

Authors:
\itemize{
  \item Stefan Kloppenborg \email{stefan@kloppenborg.ca}
}

}
\keyword{internal}
