% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fff_shear}
\alias{fff_shear}
\title{Example shear stress-shear strain data}
\format{
\subsection{\code{fff_shear}}{

A data frame with 2,316 rows and 3 columns:
\describe{
\item{Coupon}{the coupon ID}
\item{Stress}{the shear stress measurement \verb{[psi]}}
\item{Strain}{the shear strain measurement \verb{[in/in]}}
}
}
}
\usage{
fff_shear
}
\description{
Example shear stress-strain data. This data was collected using a novel
shear test method. Coupons were made using a thermoset via
Fused Filament Fabrication (FFF).
This data requires some clean-up, including removal of the "toe", offsetting
the strain, and removal of the post-failure data points. These procedures
are demonstrated in the \code{stress-strain} vignette. See:
\code{vignette("stress-strain", package = "cmstatrExt")}
}
\keyword{datasets}
