% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k_equiv_one_sample}
\alias{k_equiv_one_sample}
\title{Calculate the factors for a one-sample acceptance test}
\usage{
k_equiv_one_sample(alpha, m)
}
\arguments{
\item{alpha}{the nominal significance of the test}

\item{m}{the size of the acceptance sample}
}
\value{
A vector of length 2 with the contents \code{c(k1, k2)}
}
\description{
Calculates the factors k1 and k2, which are used for setting acceptance
values for lot acceptance. These factors consider
the size of acceptance sample (\code{m}).
}
\details{
This function is equivalent to \code{\link[cmstatr:k_equiv]{cmstatr::k_equiv()}}, but is implemented
in C++ instead of in R, and is hence slightly faster.
}
\references{
Vangel, M. (2002). Lot Acceptance and Compliance Testing Using the
Sample Mean and an Extremum.
Technometrics. https://doi.org/10.1198/004017002188618428
}
