% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k_equiv_two_sample}
\alias{k_equiv_two_sample}
\title{Calculate the factors for a two-sample acceptance test}
\usage{
k_equiv_two_sample(n, m, alpha)
}
\description{
Calculate the factors for a two-sample acceptance test
}
