% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{p_equiv}
\alias{p_equiv}
\title{p-Value for one-sample equivalency}
\usage{
p_equiv(m, t1, t2)
}
\arguments{
\item{m}{the size of the acceptance sample}

\item{t1}{the test statistic described above. May be a vector.}

\item{t2}{the test statistic described above. May be a vector.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=p_equiv_one_sample]{p_equiv_one_sample()}} instead of \code{p_equiv()}.
}
