% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-fit.R
\name{print.average_curve}
\alias{print.average_curve}
\title{Print an \code{average_curve} object}
\usage{
\method{print}{average_curve}(x, ...)
}
\arguments{
\item{x}{an \code{average_curve} object}

\item{...}{additional arguments passed to \code{print.lm}}
}
\value{
The object passed to this method is invisibly returned
(via \code{invisible(x)}).
}
\description{
Print an \code{average_curve} object
}
\seealso{
\code{\link[=average_curve]{average_curve()}}
}
