% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-fit.R
\name{print.average_curve_optim}
\alias{print.average_curve_optim}
\title{Print an \code{average_curve_optim} object}
\usage{
\method{print}{average_curve_optim}(x, ...)
}
\arguments{
\item{x}{an \code{average_curve_optim} object}

\item{...}{not used}
}
\value{
The object passed to this method is invisibly returned
(via \code{invisible(x)}).
}
\description{
Print an \code{average_curve_optim} object
}
\seealso{
\code{\link[=average_curve_optim]{average_curve_optim()}}
}
