% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-fit.R
\name{summary.average_curve}
\alias{summary.average_curve}
\title{Summarize an \code{average_curve} object}
\usage{
\method{summary}{average_curve}(object, ...)
}
\arguments{
\item{object}{an \code{average_curve} object}

\item{...}{arguments passed to \code{summary.lm}}
}
\value{
No return value. This method only produces printed output.
}
\description{
Summarize an \code{average_curve} object
}
\seealso{
\code{\link[=average_curve]{average_curve()}}
}
