% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-fit.R
\name{summary.average_curve_lm}
\alias{summary.average_curve_lm}
\title{Summarize an \code{average_curve_lm} object}
\usage{
\method{summary}{average_curve_lm}(object, ...)
}
\arguments{
\item{object}{an \code{average_curve_lm} object}

\item{...}{arguments passed to \code{summary.lm}}
}
\value{
No return value. This method only produces printed output.
}
\description{
Summarize an \code{average_curve_lm} object
}
\seealso{
\code{\link[=average_curve_lm]{average_curve_lm()}}
}
